TLRC-895: Stage 2 validations, TLRC-1010, TLRC-1015: Csv service implementation.

+++ src/Sfa.Tl.ResultsAndCertification.Models/Contracts/BulkRegistrationResponse.cs
@@ -4,7 +4,7 @@
     {
         public bool IsSuccess { get; set; }
         public string BlobErrorFileName { get; set; }
-        public long ErrorFileSize { get; set; }
+        public double ErrorFileSize { get; set; }
         public Stats Stats { get; set; }
     }
